requirements.txt
Django==1.9.8
djangorestframework==3.6.4
djangorestframework-xml==1.3.0
django-filter==1.0.4
